ml/touch/requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
joblib>=1.3
torch>=2.2
onnx>=1.15
onnxruntime>=1.18
onnx2tf==1.19.0
tensorflow==2.14.1
tf-keras==2.14.0
psutil>=5.9